requirements.txt
alembic==1.0.7
astroid==2.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Migrate==2.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
future==0.17.1
gunicorn==19.9.0
idna==2.8
isort==4.3.5
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.0.1
psycopg2==2.7.7
pylint==2.2.2
python-dateutil==2.8.0
python-editor==1.0.4
python-twitter==3.5
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
SQLAlchemy==1.2.18
typed-ast==1.3.1
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.11.1
